Refactor parsing for JSON lines and schema validation
This should have been two separate commits. It removes problematic
instances of JSON_PRETTY_PRINT which was breaking parsing for JSON
Lines. It also updates the use of the JSON Schema parsing library
(specifically reference resolution) and removes the local copy of the
library to solely rely on the one from composer.