Deploy to cloud.gov with consistent PHP version
The deploy to cloud.gov has the message:

```
   WARNING: A version of PHP has been specified in both `composer.json` and `./bp-config/options.json`.
   WARNING: The version defined in `composer.json` will be used.
   Installing PHP
 ```

and prior to this change would then note that PHP '>= 5.4.0' would
resolve to the default PHP version: 7.2.X.

by specifying a PHP version in the ~7.1 class we get a consistent
release with cloud.gov using the 7.1 latest: 7.1.29.

I've also noted that we don't use phpunit or json-schema-test-suite
so I've pulled those requirements.

Lastly, for docker, bin/composer needed a tweak to run w/o a defined
ENTRYPOINT, our Docker image needs `zip`