Almost works runs on cloud.gov except base_url
There was an earlier commit that claimed this ran on cloud.gov
but we've pulled in some stuff for back-compatibility to Ansible
that seems to have broken the ability to find the Composer `vendor/`
directory.

The docs claim that `.bp-config/options.json` can contain something like

```javascript
{
    "COMPOSER_VENDOR_DIR": "/home/vcap/app/lib/vendor"
}

```

but that results in no vendor files, I think because the composer files
are installed before the code checkout.

But the `config.php` and `database.php` files were eaching doing their
own `require_once` of `autoload.php` and that only needs to happen
in `index.php` so I'v removed those.

Now we need to fix base_url.