Generate lock file for PHP 7.0
Since our composer.lock file is generated with PHP 7.1, some dependencies were
pulled in that are only compatible with PHP 7.1+. This isn't compatible with
the BSP environment where we are running PHP 7.0. We have to generate the
composer.lock with the same PHP version across environments and update them all
in tandem (or implement work-arounds).