Clean up terminal and cron output
Log everything but actual errors as debug information in the logs. For actual errors, log them as errors. We also still echo errors for the benefit of terminal or cron observers! It's just that they're the only output shown, so that anomalies stand out when auditing the output of a run that you expect to succeed.